fix: Extensions contained in `PactVerificationService`
